feat(sdk): move the normalizer and AST access shapes into the SDK
The flow URL/method normalizer (normalizePath, normalizeMethod,
bindingKey, the catch-all helpers, the HttpMethod/ServedMethod
vocabulary) moves verbatim to packages/dxkit-sdk/src/http-normalize.ts.
There is ONE normalizer, and extensions get the exact function dxkit
runs — a second implementation on either side of the extension
boundary is the drift class Rule 2 exists to kill.

ParsedFile, walk, and the Node/Tree re-exports move to the SDK's
ast.ts. ParsedFile becomes generic over the language id (extensions
see `string`; dxkit narrows to its own pack union — the union itself
is not frozen surface). ParseFileFn/ParseSourceFn freeze the host
parser signatures the plugin runtime will bind. src/ast/parse.ts and
src/analyzers/flow/normalize.ts stay as re-export bridges, so no
consumer changes an import.
